chore: remove outdated configuration classes and consolidate service beans
- Deleted redundant configuration classes for `GameService`, `UserService`, `SearchService`, and related components.
- Merged bean definitions into `ApplicationServiceConfiguration` for a simplified and centralized setup.
- Updated imports and package structure for consistency with refactored service layouts.
