--- lsadf_core/src/main/java/com/lsadf/core/application/ApplicationServiceConfiguration.java
@@ -1,33 +1,0 @@
-/*
- * Copyright © 2024-2025 LSDAF
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.lsadf.core.application;
-
-import com.lsadf.core.application.game.GameServiceConfiguration;
-import com.lsadf.core.application.info.config.GlobalInfoConfiguration;
-import com.lsadf.core.application.search.config.SearchConfiguration;
-import com.lsadf.core.application.user.config.UserConfiguration;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-/** Configuration class for the services. */
-@Configuration
-@Import({
-  GameServiceConfiguration.class,
-  SearchConfiguration.class,
-  UserConfiguration.class,
-  GlobalInfoConfiguration.class
-})
-public class ApplicationServiceConfiguration {}
