--- lsadf_core/src/main/java/com/lsadf/core/application/game/save/currency/config/CurrencyConfiguration.java
@@ -1,43 +1,0 @@
-/*
- * Copyright © 2024-2025 LSDAF
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.lsadf.core.application.game.save.currency.config;
-
-import com.lsadf.core.application.cache.CacheManager;
-import com.lsadf.core.application.game.save.currency.CurrencyCachePort;
-import com.lsadf.core.application.game.save.currency.CurrencyRepositoryPort;
-import com.lsadf.core.application.game.save.currency.CurrencyService;
-import com.lsadf.core.application.game.save.currency.impl.CurrencyServiceImpl;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * CurrencyConfiguration is a configuration class that defines the bean creation for
- * CurrencyService. It acts as a central configuration for providing the necessary dependencies
- * required by the CurrencyService implementation.
- */
-@Configuration
-public class CurrencyConfiguration {
-
-  public static final String CURRENCY_CACHE = "currencyCache";
-
-  @Bean
-  public CurrencyService currencyService(
-      CacheManager cacheManager,
-      CurrencyRepositoryPort currencyRepositoryPort,
-      CurrencyCachePort currencyCache) {
-    return new CurrencyServiceImpl(cacheManager, currencyRepositoryPort, currencyCache);
-  }
-}
