--- lsadf_core/src/main/java/com/lsadf/core/application/game/save/metadata/config/GameMetadataConfiguration.java
@@ -1,37 +1,0 @@
-/*
- * Copyright © 2024-2025 LSDAF
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.lsadf.core.application.game.save.metadata.config;
-
-import com.lsadf.core.application.cache.CacheManager;
-import com.lsadf.core.application.game.save.metadata.GameMetadataCachePort;
-import com.lsadf.core.application.game.save.metadata.GameMetadataRepositoryPort;
-import com.lsadf.core.application.game.save.metadata.GameMetadataService;
-import com.lsadf.core.application.game.save.metadata.impl.GameMetadataServiceImpl;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class GameMetadataConfiguration {
-  @Bean
-  public GameMetadataService gameMetadataService(
-      CacheManager cacheManager,
-      GameMetadataRepositoryPort gameMetadataRepositoryPort,
-      GameMetadataCachePort gameMetadataCachePort) {
-    return new GameMetadataServiceImpl(
-        cacheManager, gameMetadataRepositoryPort, gameMetadataCachePort);
-  }
-}
