+++ lsadf_core/src/test/java/com/lsadf/core/unit/application/game/save/CharacteristicsServiceTests.java
@@ -23,15 +23,14 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
 import com.lsadf.core.application.cache.CacheManager;
+import com.lsadf.core.application.game.save.characteristics.CharacteristicsCachePort;
 import com.lsadf.core.application.game.save.characteristics.CharacteristicsRepositoryPort;
 import com.lsadf.core.application.game.save.characteristics.CharacteristicsService;
 import com.lsadf.core.application.game.save.characteristics.impl.CharacteristicsServiceImpl;
-import com.lsadf.core.application.shared.CachePort;
 import com.lsadf.core.domain.game.save.characteristics.Characteristics;
 import com.lsadf.core.exception.http.NotFoundException;
 import java.util.Optional;
 import java.util.UUID;
-import org.jspecify.annotations.NonNull;
 import org.junit.jupiter.api.*;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
@@ -42,7 +41,7 @@ class CharacteristicsServiceTests {
 
   @Mock private CharacteristicsRepositoryPort characteristicsRepositoryPort;
   @Mock private CacheManager cacheManager;
-  @Mock private CachePort<@NonNull Characteristics> characteristicsCache;
+  @Mock private CharacteristicsCachePort characteristicsCache;
 
   private static final UUID UUID = java.util.UUID.randomUUID();
